% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balancoPeriodoClp.R
\name{balancoPeriodoClp}
\alias{balancoPeriodoClp}
\title{Calcula um balanco de potencia usando o solver Coin}
\usage{
balancoPeriodoClp(
  periodo,
  idSerieHidro,
  balancoResumido,
  conexao,
  df.custoDefict,
  df.geracaoTermicaTotal,
  df.geracaoTermicaGnlTotal,
  df.geracaoTransmissaoTotal,
  df.geracaoRenovaveisTotal,
  df.geracaoArmazenamentoTotal,
  df.limitesAgrupamentoLinhasTotal,
  df.demanda,
  df.geracaoHidroTotal,
  df.agrupamentoLinhas,
  tipoCaso,
  numeroCaso,
  codModelo,
  df.subsistemas,
  cvuHidro,
  df.defictRealocadoTotal
)
}
\arguments{
\item{periodo}{vetor com o anoMes a ser ser processado. Ex. 201805}

\item{idSerieHidro}{variavel com o valor da serie hidrologica para a qual sera calculada o balanco}

\item{balancoResumido}{variavel binaria para decidir se vai calcular somente o balanco resumido (\code{BPO_A16_BALANCO}) ou 
tambem o por gerador (\code{BPO_A17_BALANCO_GERADOR})}

\item{conexao}{conexao com o banco de dados (classe SQLiteConnection)}

\item{df.custoDefict}{data frame com custos de deficit}

\item{df.geracaoTermicaTotal}{data frame com os dados das geracoes termicas}

\item{df.geracaoTermicaGnlTotal}{data frame com os dados das geracoes termicas com despacho antecipado GNL}

\item{df.geracaoTransmissaoTotal}{data frame com os dados de transmissao}

\item{df.geracaoRenovaveisTotal}{data frame com os dados de renovaveis}

\item{df.geracaoArmazenamentoTotal}{data frame com os dados de projetos de armazenamento}

\item{df.limitesAgrupamentoLinhasTotal}{data frame com os dados de agrupamento de transmissao}

\item{df.demanda}{data frame com os dados de demanda}

\item{df.geracaoHidroTotal}{data frame com os dados de hidro}

\item{df.agrupamentoLinhas}{data frame com os dados de agrupamento de linhas}

\item{tipoCaso}{variavel com o tipo de caso}

\item{numeroCaso}{variavel com o numero do caso}

\item{codModelo}{variavel com o codigo do modelo}

\item{df.subsistemas}{data frame com os dados de subsistemas}

\item{cvuHidro}{valor de CVU das hidro}

\item{df.defictRealocadoTotal}{data frame com os dados do limite de deficit a ser realocado por subsistema}
}
\value{
\code{lt.resultado} lista com data frames com as estruturas das tabelas com os resultados dos balancos. 
A lista pode ter 2 ou 3 data frames conforme definido pela variavel \code{balancoResumido}.
\itemize{
\item Para \code{balancoResumido} == \code{F}: lista com 3 data frames com as estruturas das tabelas \code{BPO_A16_BALANCO} (\code{df.resultado}),
\code{BPO_A17_BALANCO_GERADOR} (\code{df.resultadoGerador}) e \code{BPO_A20_BALANCO_SUBSISTEMA} (\code{df.resultadoCMO}) para cada periodo (anoMes)
\item Para \code{balancoResumido} == \code{T}: lista com 2 data frames com as estruturas das tabelas \code{BPO_A16_BALANCO} (\code{df.resultado}) e
\code{BPO_A20_BALANCO_SUBSISTEMA} (\code{df.resultadoCMO})
}
}
\description{
Monta e resolve o problema linear de um balanco de potencia usando Clp (Coin-or linear programming). 
Funcao criada para poder executar processamento paralelo.
}
\examples{
\dontrun{
balancoPeriodoClp(201901, 1, T, conexao, df.custoDefict, df.geracaoTermicaTotal,
df.geracaoTermicaGnlTotal, df.geracaoTransmissaoTotal, df.geracaoRenovaveisTotal,
df.geracaoArmazenamentoTotal, df.limitesAgrupamentoLinhasTotal, df.demanda,
df.geracaoHidroTotal,df.agrupamentoLinhas, tipoCaso, numeroCaso, codModelo,
df.subsistemas, cvuHidro, df.defictRealocadoTotal)
}

}
