% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculaBalancoParalelo.R
\name{calculaBalancoParalelo}
\alias{calculaBalancoParalelo}
\title{Calcula todo o processo de balanco de potencia}
\usage{
calculaBalancoParalelo(
  baseSQLite,
  tipoCaso,
  numeroCaso,
  codModelo,
  cvuTransmissao,
  cvuHidro,
  cvuRenovaveis,
  cvuOutrasTermicas,
  balancoResumido = T,
  distribuicaoDeficit = 0.05
)
}
\arguments{
\item{baseSQLite}{vetor de caracteres com nome e localizacao da base de dados. Ex: "bp.slqlite3"}

\item{tipoCaso}{variavel com o tipo de caso}

\item{numeroCaso}{variavel com o numero do caso}

\item{codModelo}{variavel com o codigo do modelo}

\item{cvuTransmissao}{valor de "CVU" para uso da transmissao. Valor declarado baixo para dar prioridade. Ex. 1e-10}

\item{cvuHidro}{valor de "CVU" para uso das hidraulicas. Valor declarado baixo para dar 3a prioridade. Ex. 2e-8}

\item{cvuRenovaveis}{valor de "CVU" para uso das outras renovaveis. Valor declarado baixo para dar 2a prioridade. Ex. 1e-8}

\item{cvuOutrasTermicas}{valor de "CVU" para uso das termicas com CVU decalaro de 0. Valor declarado baixo para dar 4a prioridade. Ex. 0.1}

\item{balancoResumido}{variavel binaria para decidir se vai gravar somente o balanco resumido (\code{BPO_A16_BALANCO}) ou 
tambem o por gerador (\code{BPO_A17_BALANCO_GERADOR}). Valor padrao T.}

\item{distribuicaoDeficit}{variavel com valor (percentual) da demanda para ser o limite de disponibilidade do deficit distribuido. Valor padrao 0.05}
}
\value{
\code{mensagem} vetor de caracteres com a mensagem de sucesso de gravacao dos resultados dos balancos 
nas tabelas BPO_A16_BALANCO, BPO_A17_BALANCO_GERADOR e BPO_A20_BALANCO_SUBSISTEMA
}
\description{
Monta e resolve o problema linear de um balanco de potencia. Funcao criada para poder executar processamento paralelo.
}
\examples{
\dontrun{
calculaBalancoParalelo("C:/PDE2027_Caso080/bp.slqlite3", 1, 80, 1, 1e-10, 2e-8, 1e-8, 0.1, F)}

}
