% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculaDisponibilidadeHidro.R
\name{calculaDisponibilidadeHidro}
\alias{calculaDisponibilidadeHidro}
\title{Calcula a disponibilidade hidraulica para o Balanco de Potencia}
\usage{
calculaDisponibilidadeHidro(
  baseSQLite,
  pastaCaso,
  pastaSaidas,
  tipoCaso,
  numeroCaso,
  codModelo,
  lt.dadosTucurui,
  flagVert = FALSE,
  flagUHE = FALSE,
  execShiny = FALSE,
  tipoModulacao
)
}
\arguments{
\item{baseSQLite}{nome e localizacao da base SQLite do balanco de ponta}

\item{pastaCaso}{localizacao da pasta com os arquivos do NEWAVE do caso a ser analisado no balanco de ponta}

\item{pastaSaidas}{localizacao dos arquivos de saida do modulo NWLISTOP}

\item{tipoCaso}{valor inteiro. 1:PDE; 2:PMO e 3;Garantia Fisica}

\item{numeroCaso}{valor inteiro com o numero do caso}

\item{codModelo}{valor inteiro com o codigo do modelo. 1:NEWAVE}

\item{lt.dadosTucurui}{lista com dados referentes a UHE Tucurui para calculo do PDisp}

\item{flagVert}{booleano que indica se considera ou nao o vertimento para todas as UHE}

\item{flagUHE}{booleano que indica se as saidas individuais por UHE deverao ou nao ser gravadas na base de dados}

\item{execShiny}{booleano que indica se a funcao esta sendo executada em um contexto reativo, para atualizacao da barra de progresso}

\item{tipoModulacao}{valor inteiro. 1:Modulacao por REE; 2:Modulacao por UHE}
}
\value{
\code{mensagem} vetor de caracteres com a mensagem de sucesso de gravacao na base
}
\description{
Faz os calculos da disponibilidade hidraulica
}
