% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculaDisponibilidadeTabela.R
\name{calculaDisponibilidadeTabela}
\alias{calculaDisponibilidadeTabela}
\title{Calcula a disponibilidade hidraulica para UHEs que modulam segundo tabela defluencia x potencia disponivel}
\usage{
calculaDisponibilidadeTabela(
  tipoCaso,
  numeroCaso,
  codModelo,
  pastaCaso,
  UHE,
  df.saidasHidro,
  lt.hidrogramaBM,
  df.hidrograma,
  df.tabelaModulacao,
  flagVert
)
}
\arguments{
\item{tipoCaso}{valor inteiro. 1:PDE; 2:PMO e 3;Garantia Fisica}

\item{numeroCaso}{valor inteiro com o numero do caso}

\item{codModelo}{valor inteiro com o codigo do modelo. 1:NEWAVE}

\item{pastaCaso}{caracter com a localizacao dos arquivos NEWAVE e auxliares do BP}

\item{UHE}{vetor com o codigo das UHE que modulam por tabela}

\item{df.saidasHidro}{data frame com dados da tabela BPO_A06_SAIDA_HIDRO_NEWAVE}

\item{lt.hidrogramaBM}{lista contendo os dados do hidrograma de Belo Monte}

\item{df.hidrograma}{data frame com dados do hidrograma de Belo Monte}

\item{df.tabelaModulacao}{data frame com dados de defluencia x potencia disponivel}

\item{flagVert}{booleano que indica se considera ou nao o vertimento para todas as UHE}
}
\value{
\code{df.dadosUHEModulamTabela} data frame com dados de disponibilidade individuais das UHE que modulam conforme tabela
}
\description{
Calcula a disponibilidade hidraulica para UHEs que modulam segundo tabela defluencia x potencia disponivel
}
