% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculaDisponibilidadeTipo1.R
\name{calculaDisponibilidadeTipo1}
\alias{calculaDisponibilidadeTipo1}
\title{Calcula a disponibilidade hidraulica usando a metodologia da EPE}
\usage{
calculaDisponibilidadeTipo1(
  tipoCaso,
  numeroCaso,
  codModelo,
  df.saidasHidro,
  df.dadosUHE,
  df.dadosVigentes,
  df.dadosMaquinasUHE,
  df.potMaquinas,
  df.dadosCaso,
  codTucurui,
  flagVert
)
}
\arguments{
\item{tipoCaso}{valor inteiro. 1:PDE; 2:PMO e 3;Garantia Fisica}

\item{numeroCaso}{valor inteiro com o numero do caso}

\item{codModelo}{valor inteiro com o codigo do modelo. 1:NEWAVE}

\item{df.saidasHidro}{data frame com dados da tabela BPO_A06_SAIDA_HIDRO_NEWAVE}

\item{df.dadosUHE}{data frame com dados da tabela BPO_A03_DADOS_UHE}

\item{df.dadosVigentes}{data frame com dados da tabela BPO_A05_DADOS_VIGENTES_UHE}

\item{df.dadosMaquinasUHE}{data frame com dados da tabela BPO_A04_MAQUINAS_UHE}

\item{df.potMaquinas}{data frame com dados de potencia das UHE}

\item{df.dadosCaso}{data frame com dados do caso em execucao}

\item{codTucurui}{codigo atribuido para a usina de Tucurui}

\item{flagVert}{booleano que indica se considera ou nao o vertimento para todas as UHE}
}
\value{
\code{df.dadosCalculadosUHE} data frame com dados de disponibilidade individuais das UHE
}
\description{
Calcula a disponibilidade hidraulica usando a metodologia da EPE
}
