% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criaBDBalanco.R
\name{criaBDBalanco}
\alias{criaBDBalanco}
\title{Cria banco de dados do balanco de potencia}
\usage{
criaBDBalanco(caminho, nomeBD)
}
\arguments{
\item{caminho}{localizacao do banco de dados}

\item{nomeBD}{nome do banco de dados}
}
\value{
mensagem de sucesso
}
\description{
Executa script SQL e cria banco de dados SQLite com estrutura dos dados do balanco de potencia
}
\examples{
criaBDBalanco("C:/PDE2027_Caso080/BD", "bd_balanco_pde")

}
