% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvar.R
\name{cvar}
\alias{cvar}
\title{Calcula conditional-value-at-risk (CVaR)}
\usage{
cvar(x, cvar)
}
\arguments{
\item{x}{vetor com valores a se calcular o CVaR}

\item{cvar}{valor de cvar aplicado. Ex. 0.01, 0.02, 0.05}
}
\value{
valor de cvar aplicado no vetor x
}
\description{
Cria a funcao que calcula o valor de risco condicional (CVaR)
}
