% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadosGraficosGF.R
\name{dadosGraficosGF}
\alias{dadosGraficosGF}
\title{Exibe graficos de CVaR com criterios de GF}
\usage{
dadosGraficosGF(baseSQLite, tipoCaso, numeroCaso, codModelo, tipoGrafico)
}
\arguments{
\item{baseSQLite}{nome e localizacao da base SQLite do balanco de potencia}

\item{tipoCaso}{valor inteiro. 1:PDE; 2:PMO e 3;Garantia Fisica}

\item{numeroCaso}{valor inteiro com o numero do caso}

\item{codModelo}{valor inteiro com o codigo do modelo. 1:NEWAVE}

\item{tipoGrafico}{valor numerico identificando o tipo de grafico. 10: CVAR mensal; 11: VAR mensal; 12: VAR anual}
}
\value{
objeto do tipo plotly
}
\description{
Monta graficos de Profundidade de Deficit por CVAR de um caso especifico com criterios de GF
}
