% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadosGraficosVaR.R
\name{dadosGraficoVAR}
\alias{dadosGraficoVAR}
\title{Exporta os dados dos graficos de VaR}
\usage{
dadosGraficoVAR(
  baseSQLite,
  tipoCaso,
  numeroCaso,
  codModelo,
  inicioHorizonteGrafico,
  fimHorizonteGrafico,
  tipoGrafico
)
}
\arguments{
\item{baseSQLite}{nome e localizacao da base SQLite do balanco de potencia}

\item{tipoCaso}{valor inteiro. 1:PDE; 2:PMO e 3;Garantia Fisica}

\item{numeroCaso}{valor inteiro com o numero do caso}

\item{codModelo}{valor inteiro com o codigo do modelo. 1:NEWAVE}

\item{inicioHorizonteGrafico}{valor numerico do ano de inicio do horizonte para o grafico. Formato: AAAA. Ex: 2020}

\item{fimHorizonteGrafico}{valor numerico do ano de fim do horizonte para o grafico. Formato: AAAA. Ex:2029}

\item{tipoGrafico}{valor numerico identificando o tipo de grafico. 1: Mensal por patamar; 2: Mensal em linha; 3:Anual}
}
\value{
tib.dadosGraficoVAR tibble com os dados do grafico de CVaR
}
\description{
Exporta os dados dos graficos de Profundidade de Deficit por VAR de um caso especifico
}
