% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficoVAR.R
\name{graficoVAR}
\alias{graficoVAR}
\title{Exibe graficos de VaR}
\usage{
graficoVAR(
  baseSQLite,
  tipoCaso,
  numeroCaso,
  codModelo,
  inicioHorizonteGrafico,
  fimHorizonteGrafico,
  tipoGrafico,
  tituloGraficoVARMes = paste0("Profundidade de Déficit - VaR Mensal - Caso ",
    numeroCaso),
  tituloGraficoVARAno = paste0("Profundidade de Déficit - VaR Anual - Caso ",
    numeroCaso)
)
}
\arguments{
\item{baseSQLite}{nome e localizacao da base SQLite do balanco de potencia}

\item{tipoCaso}{valor inteiro. 1:PDE; 2:PMO e 3;Garantia Fisica}

\item{numeroCaso}{valor inteiro com o numero do caso}

\item{codModelo}{valor inteiro com o codigo do modelo. 1:NEWAVE}

\item{inicioHorizonteGrafico}{valor numerico do ano de inicio do horizonte para o grafico. Formato: AAAA. Ex: 2020}

\item{fimHorizonteGrafico}{valor numerico do ano de fim do horizonte para o grafico. Formato: AAAA. Ex:2029}

\item{tipoGrafico}{valor numerico identificando o tipo de grafico. 5: Mensal por patamar; 6: Mensal em linha; 7:Anual}

\item{tituloGraficoVARMes}{vetor de caracteres com o titulo do grafico de VAR mensal - Nao obrigatorio - valor padrao com numero do caso}

\item{tituloGraficoVARAno}{vetor de caracteres com o titulo do grafico de VAR anual - Nao obrigatorio - valor padrao com numero do caso}
}
\value{
objeto do tipo plotly
}
\description{
Monta graficos de Profundidade de Deficit por VAR de um caso especifico
}
