% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gravaDadosDisponibilidadeOutrasFontesRenovaveisBDBP.R
\name{gravacaoDadosDisponibilidadeOutrasFontesBDBP}
\alias{gravacaoDadosDisponibilidadeOutrasFontesBDBP}
\title{Gravacao dos dados de disponilidade das outras fontes renovaveis}
\usage{
gravacaoDadosDisponibilidadeOutrasFontesBDBP(
  pastaCaso,
  conexao,
  tipoCaso,
  numeroCaso,
  tipoDemanda,
  codModelo,
  anoMesInicioMDI,
  anoMesFimMDI
)
}
\arguments{
\item{pastaCaso}{caracter com a localizacao dos arquivos NEWAVE e auxliares do BP.}

\item{conexao}{caracter com a conexao com o Banco de Dados do Balanco de Potencia.}

\item{tipoCaso}{caracter com o tipo de caso simulado. [1]=PDE [2]=PMO [3]=GF.}

\item{numeroCaso}{caracter com o numero do caso, definido pelo usuario.}

\item{tipoDemanda}{caracter com a definicao do tipo de demanda do caso. [1]=Deterministica [2]=Liquida}

\item{codModelo}{valor inteiro com o codigo do modelo. 1:NEWAVE}

\item{anoMesInicioMDI}{caracter com o ano e mes do inicio da simulacao do MDI.}

\item{anoMesFimMDI}{caracter com o ano e mes do final da simulacao do MDI.}
}
\value{
\code{lista} 
\itemize{
\item \code{mensagem} vetor de caracteres com a mensagem de sucesso de gravacao nas tabelas BPO_A13_DISPONIBILIDADE_OFR, 
BPO_A18_TIPOS_OFR e BPO_A19_FATOR_PONTA_OFR 
\item \code{df.capacidadeOFR} data frame com a capacidade instalada das renovaveis
}
}
\description{
Faz a gravacao dos dados de disponibilidade das outras fontes renovaveis do NEWAVE no banco de dados do Balanco de Potencia (BDBP)
Os dados sao gravados na tabela BPO_A13_DISPONIBILIDADE_OFR do BDBP. Alem disso, grava as tabelas de apoio BPO_A18_TIPOS_OFR e BPO_A19_FATOR_PONTA_OFR
}
\examples{
\dontrun{
gravacaoDadosDisponibilidadeOutrasFontesBDBP("C:/PDE2027_Caso080", conexao, 
1, 80, 1, 1, 201901, 203312)
}

}
