% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravaDadosLimitesAgrupIntercambioBDBP.R
\name{gravacaoDadosLimitesAgrupIntercambioBDBP}
\alias{gravacaoDadosLimitesAgrupIntercambioBDBP}
\title{Gravacao dos dados dos limites dos agrupamentos de intercambios ao longo do horizonte de simulacao}
\usage{
gravacaoDadosLimitesAgrupIntercambioBDBP(
  pastaCaso,
  conexao,
  tipoCaso,
  numeroCaso,
  codModelo
)
}
\arguments{
\item{pastaCaso}{caracter com a localizacao dos arquivos NEWAVE e auxliares do BP.}

\item{conexao}{caracter com a conexao com o Banco de Dados do Balanco de Potencia.}

\item{tipoCaso}{caracter com o tipo de caso simulado. [1]=PDE [2]=PMO [3]=GF.}

\item{numeroCaso}{caracter com o numero do caso, definido pelo usuario.}

\item{codModelo}{valor inteiro com o codigo do modelo. 1:NEWAVE}
}
\value{
\code{mensagem} vetor de caracteres com a mensagem de sucesso de gravacao na tabela BPO_A12_LIMITE_AGRUPAMENTOS_INTERCAMBIO
}
\description{
Faz a gravacao dos dados dos limites dos agrupamento de intercambios ao longo do horizonte de simulacao do NEWAVE no banco de dados do Balanco de Potencia (BDBP)
Os dados sao gravados na tabela BPO_A12_LIMITE_AGRUPAMENTOS_INTERCAMBIO do BDBP.
}
\examples{
\dontrun{
gravacaoDadosLimitesAgrupIntercambioBDBP("C:/PDE2027_Caso080", conexao, 1, 80, 1)
}

}
