% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravaDadosReservaBDBP.R
\name{gravacaoDadosReservaBDBP}
\alias{gravacaoDadosReservaBDBP}
\title{Gravacao dos dados de reserva operativa}
\usage{
gravacaoDadosReservaBDBP(
  pastaCaso,
  conexao,
  tipoCaso,
  numeroCaso,
  codModelo,
  df.energiaOFR
)
}
\arguments{
\item{pastaCaso}{caracter com a localizacao dos arquivos NEWAVE e auxliares do BP.}

\item{conexao}{caracter com a conexao com o Banco de Dados do Balanco de Potencia.}

\item{tipoCaso}{caracter com o tipo de caso simulado. [1]=PDE [2]=PMO [3]=GF.}

\item{numeroCaso}{caracter com o numero do caso, definido pelo usuario.}

\item{codModelo}{valor inteiro com o codigo do modelo. 1:NEWAVE}

\item{df.energiaOFR}{data frame com a energia esperada mensal das renovaveis}
}
\value{
\code{mensagem} vetor de caracteres com a mensagem de sucesso de gravacao na tabela BPO_A21_RESERVA
}
\description{
Faz a gravacao dos dados de de reserva operativa no banco de dados do Balanco de Potencia (BDBP).
Os dados sao gravados na tabela BPO_A21_RESERVA do BDBP. Consulta as tabelas BPO_A13_DISPONIBILIDADE_OFR e BPO_A10_DEMANDA.
}
\examples{
\dontrun{
gravacaoDadosReservaBDBP("C:/PDE2027_Caso080", conexao, 1, 80, 1, df.energiaOFR)
}

}
