% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravaSaidasNewaveBDBP.R
\name{gravaSaidasNewaveBDBP}
\alias{gravaSaidasNewaveBDBP}
\title{Grava dados de armazenamento e geracao das usinas hidraulicas no banco de dados do balanco de potencia}
\usage{
gravaSaidasNewaveBDBP(
  pasta,
  pastaSaidas,
  conexao,
  tipoCaso,
  numeroCaso,
  codModelo
)
}
\arguments{
\item{pasta}{localizacao dos arquivos do NEWAVE}

\item{pastaSaidas}{localizacao dos arquivos de saida do modulo NWLISTOP}

\item{conexao}{conexao com o banco de dados (classe SQLiteConnection)}

\item{tipoCaso}{valor inteiro. 1:PDE; 2:PMO e 3;Garantia Fisica}

\item{numeroCaso}{valor inteiro com o numero do caso}

\item{codModelo}{valor inteiro com o codigo do modelo. 1:NEWAVE}
}
\value{
\code{mensagem} vetor de caracteres com a mensagem de sucesso de gravacao na tabela BPO_A06_SAIDA_HIDRO_NEWAVE
}
\description{
Grava dados de armazenamento, vertimento turbinavel e geracao das usinas hidraulicas originalmente nos 
arquivos earmfpXXX.out, verturbxxx.* e ghtotxxx.out na tabela BPO_A06_SAIDA_HIDRO_NEWAVE 
do banco de dados do balanco de potencia. Usa funcoes do pacote (\code{leitorrmpe}).
}
\examples{
\dontrun{
gravaSaidasNewaveBDBP("C:/PDE2027_Caso080", "C:/PDE2027_Caso080/nwlistop", conexao, 1, 80, 1)
}

}
