% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitorDadosGerais.R
\name{leituraDadosGerais}
\alias{leituraDadosGerais}
\title{Leitor dos dados gerais}
\usage{
leituraDadosGerais(arquivo)
}
\arguments{
\item{arquivo}{nome e localizacao do arquivo do NEWAVE com dados gerais}
}
\value{
\code{df.dadosGerais} data frame com informacao sobre o estudo
\itemize{
\item numero de anos de duracao do estudo(\code{$duracaoEstudo})
\item mes do inicio do estudo (\code{$mesInicio})
\item ano do inicio do estudo (\code{$anoInicio})
\item numero de anos iniciais para fins de estabilizacao no calculo da politica (\code{$anosPre})
\item numero de anos finais para fins de estabilizacao no calculo da politica (\code{$anosPos})
\item numero de series sinteticas (\code{$seriesSinteticas})
\item valor definindo tipo de simuacao final (0 nao simula; 1 serie sintetica; 2 serie historica; 3 consistencia de dados) (\code{$tipoSimulacao})
}
}
\description{
Faz a leitura do arquivo do NEWAVE com dados gerais (gder.*) e recupera dados necessarios para o balanco.
Usa como referencia para a leitura do arquivo as posicoes definidas no Manual do Usuario do
Modelo Estrategico de geracao hidrotermica a subsistemas equivalentes do Projeto NEWAVE versao 25.0.1 de setembro/2019 - paginas 17 - 25
}
\examples{
leituraDadosGerais("C:/PDE2027_Caso080/dger.d27")

}
