% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitorREE.R
\name{leituraREE}
\alias{leituraREE}
\title{Leitor dos dados dos Reservatorios Equivalentes de Energia (REE)}
\usage{
leituraREE(arquivo)
}
\arguments{
\item{arquivo}{nome e localizacao do arquivo do NEWAVE com dados de REE}
}
\value{
\code{df.ree} data frame com os dados dos REE
\itemize{
  \item codigo do REE (\code{$codREE})
  \item nome (\code{$nomeREE})
  \item codigo do sistema do REE (\code{$codSistema})
  }
}
\description{
Faz a leitura do arquivo do NEWAVE com informacao dos REE (ree.dat) e recupera o codigo, nome e sistema de cada REE cadastrada
}
\examples{
leituraREE("C:/PDE2027_Caso080/ree.dat")

}
