% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitorSistema.R
\name{leituraSistema}
\alias{leituraSistema}
\title{Leitor dos dados de sistema}
\usage{
leituraSistema(arquivo)
}
\arguments{
\item{arquivo}{nome e localizacao do arquivo do NEWAVE com dados de sistema}
}
\value{
\code{df.sistema} data frame com informacoes de sistema
\itemize{
  \item codigo do sistema (\code{$codSistema})
  \item nome do sistema (\code{$nomeSistema})
  \item tipo de sistema (\code{$tipoFicticio}) - pode assumir 1 para ficticio, ou 0 para real
  \item custo de deficit em R$/MWh (\code{$deficit})
  }
}
\description{
Faz a leitura do arquivo do NEWAVE com informacao de deficit do sistema (sistema.d**)
}
\examples{
leituraSistema("C:/PDE2027_Caso080/sistema.d27")

}
