% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraSubmotorizacaoREE.R
\name{leituraSubmotorizacaoREE}
\alias{leituraSubmotorizacaoREE}
\title{Leitor dos dados de submotorizacao por REE}
\usage{
leituraSubmotorizacaoREE(arquivo)
}
\arguments{
\item{arquivo}{nome e localizacao do arquivo do NEWAVE com dados de submotorizacao}
}
\value{
\code{df.submotorizacao} data frame com os valores de submotorizacao por REE
\itemize{
\item nome da REE (\code{$ree})
\item valor de ano e mes (\code{$anoMes})
}
}
\description{
Faz a leitura do arquivo do NEWAVE com informacao de submotorizacao por REE (relat.d**) e recupera esses valores por ano e mes
}
\examples{
leituraSubmotorizacaoREE("C:/PDE2027_Caso080/relat.d27")

}
