% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitorTabelaDadosCasos.R
\name{leituraTabelaDadosCasos}
\alias{leituraTabelaDadosCasos}
\title{Faz a leitura da tabela com dados dos casos}
\usage{
leituraTabelaDadosCasos(baseSQLite)
}
\arguments{
\item{baseSQLite}{nome e localizacao da base SQLite do balanco de ponta}
}
\value{
\code{df.dadosCaso} data frame com dados dos casos (BPO_A01_CASOS_ANALISE)
}
\description{
Faz a leitura da tabela com dados dos casos (BPO_A01_CASOS_ANALISE) e monta select input para interface grafica
}
