% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverBalanco.R
\name{serverBalanco}
\alias{serverBalanco}
\title{Servidor Shiny do Banlanco}
\usage{
serverBalanco(input, output, session)
}
\arguments{
\item{input}{entradas da camada ui}

\item{output}{saidas da camada ui}

\item{session}{objeto contendo dados e funcionalidades da da sessao}
}
\description{
Cria servidor do Balanco de Ponta
}
