% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subRBind.R
\name{subRBind}
\alias{subRBind}
\title{RBind em data frames dentro de listas}
\usage{
subRBind(lista1, lista2)
}
\arguments{
\item{lista1}{lista com data frames}

\item{lista2}{lista com data frames. Ambas as listas devem ter a mesma estrutura}
}
\value{
\code{lt.saida} lista onde os data frames de lista1 e lista2 foram agrupados (rbind) seguindo cada estrutura de data frame
}
\description{
Funcao para efetuar o rbind nos data frames dentro de listas que sao a saida da funcao balancoPeriodo()
}
\examples{
\dontrun{
subRBind(lista1, lista2)
}

}
