% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiBalancoEPE.R
\docType{data}
\name{uiBalancoEPE}
\alias{uiBalancoEPE}
\title{Interface do usuario (ui) Shiny do Balanco de Potencia}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
uiBalancoEPE
}
\description{
Cria estrutura de interface de usuario (ui) do Balanco de Potencia
}
\keyword{datasets}
