% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{var}
\alias{var}
\title{Calcula Value at Risk (VaR)}
\usage{
var(x, var)
}
\arguments{
\item{x}{vetor com valores a se calcular o VaR}

\item{var}{valor de var aplicado. Ex. 0.01, 0.02, 0.05}
}
\value{
valor de var aplicado no vetor x
}
\description{
Cria a funcao que calcula o valor de risco (VaR)
}
