% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aplicacaoBalanco.R
\name{aplicacaoBalanco}
\alias{aplicacaoBalanco}
\title{Aplicacao web do Balanco de Potencia}
\usage{
aplicacaoBalanco()
}
\description{
Executa a aplicacao web do Balanco de Potencia
}
\examples{
aplicacaoBalanco()

}
