% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carregaDadosSQLite.R
\name{carregaDadosSQLite}
\alias{carregaDadosSQLite}
\title{Carrega dados na base SQLite do Balanco de Potencia}
\usage{
carregaDadosSQLite(
  baseSQLite,
  pastaCaso,
  pastaSaidas,
  tipoCaso,
  numeroCaso,
  codModelo,
  descricaoCaso,
  horasPonta,
  idDemandaLiquida,
  idModulacao,
  sistemasNaoModulamPonta,
  sistemasNaoModulamMedia,
  sistemasModulamTabela,
  execShiny = FALSE
)
}
\arguments{
\item{baseSQLite}{nome e localizacao da base SQLite do balanco de ponta}

\item{pastaCaso}{localizacao da pasta com os arquivos do NEWAVE do caso a ser analisado no balanco de ponta}

\item{pastaSaidas}{localizacao dos arquivos de saida do modulo NWLISTOP}

\item{tipoCaso}{valor inteiro. 1:PDE; 2:PMO e 3;Garantia Fisica}

\item{numeroCaso}{valor inteiro com o numero do caso}

\item{codModelo}{valor inteiro com o codigo do modelo. 1:NEWAVE}

\item{descricaoCaso}{vetor de caracteres indicando a descricao do caso}

\item{horasPonta}{valor inteiro com o numero de horas de ponta}

\item{idDemandaLiquida}{identificador de calculo com demanda liquida. [1]=Deterministica [2]=Liquida}

\item{idModulacao}{identificador do tipo de modulacao. [1]=por REE [2]=por UHE}

\item{sistemasNaoModulamPonta}{vetor numerico com codigos dos sitemas que nao modulam na ponta}

\item{sistemasNaoModulamMedia}{vetor numerico com codigos dos sitemas que nao modulam na media}

\item{sistemasModulamTabela}{vetor numerico com codigos dos sitemas que modulam conforme tabela}

\item{execShiny}{booleano que indica se a funcao esta sendo executada em um contexto reativo, para atualizacao da barra de progresso}
}
\value{
\code{mensagem} vetor de caracteres com a mensagem de sucesso de gravacao na base
}
\description{
Faz a leitura dos arquivos do NEWAVE e carrega todos os dados necessarios na base SQLite para execucao do Balanco de Ponto.
Usa como referencia para a identificacao dos arquivos as posicoes definidas no Manual do Usuario do
Modelo Estrategico de geracao hidrotermica a subsistemas equivalentes do Projeto NEWAVE versao 25.0.1 de setembro/2019 - paginas 12 - 15
}
