% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carregaREEeSistemaSQLite.R
\name{carregaDadosSQLite}
\alias{carregaDadosSQLite}
\title{Carrega dados na base SQLite do Balanco de Ponta}
\usage{
carregaDadosSQLite(baseSQLite, pastaCaso, tipoCaso, numeroCaso, codModelo,
  descricaoCaso, horasPonta, reservaOperativa, idDemandaLiquida,
  codTucurui, cotaLimiteTucurui, geracaoLimiteTucurui, anoMesInicioMDI,
  anoMesFimMDI)
}
\arguments{
\item{arquivo}{nome e localizacao do arquivo do NEWAVE com dados de submotorizacao}
}
\value{
\code{df.submotorizacao} data frame com os valores de submotorizacao por REE
\itemize{
\item nome da REE (\code{$ree})
\item valor de ano e mes (\code{$anoMes})
}
}
\description{
Faz a leitura dos arquivos do NEWAVE e carrega todos os dados necessarios para execucao do Balanco de Ponto na sua base SQLite.
Usa como referencia para a leitura dos arquivos as posicoes definidas no Manual do Usuario do
Modelo Estrategico de geracao hidrotermica a subsistemas equivalentes do Projeto NEWAVE versao 25.0.1 de setembro/2019 - paginas 12 - 15
}
