% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carregaDadosSQLite.R
\name{carregaDadosSQLite}
\alias{carregaDadosSQLite}
\title{Carrega dados na base SQLite do Balanco de Ponta}
\usage{
carregaDadosSQLite(baseSQLite, pastaCaso, tipoCaso, numeroCaso, codModelo,
  descricaoCaso, horasPonta, reservaOperativa, idDemandaLiquida, anosPre,
  anosPos, sistemasNaoModulamPonta, sistemasNaoModulamMedia, codTucurui,
  cotaLimiteTucurui, geracaoLimiteTucurui, anoMesInicioMDI = NA,
  anoMesFimMDI = NA)
}
\arguments{
\item{baseSQLite}{nome e localizacao da base SQLite do balanco de ponta}

\item{pastaCaso}{localizacao da pasta com os arquivos do NEWAVE do caso a ser analisado no balanco de ponta}

\item{tipoCaso}{valor inteiro. 1:PDE; 2:PMO e 3;Garantia Fisica}

\item{numeroCaso}{valor inteiro com o numero do caso}

\item{codModelo}{valor inteiro com o codigo do modelo. 1:NEWAVE; 2:SUISHI}

\item{descricaoCaso}{vetor de caracteres indicando a descricao do caso}

\item{horasPonta}{valor inteiro com o numero de horas de ponta}

\item{reservaOperativa}{valor com a reserva operativa usada no calculo do balanco. Ex. 0.05}

\item{idDemandaLiquida}{identificador de calculo com demanda liquida. 1:Demanda Liquida; 0:Deterministica}

\item{anosPre}{numero de anos iniciais para fins de estabilizacao no calculo da politica}

\item{anosPos}{numero de anos finais para fins de estabilizacao no calculo da politica}

\item{sistemasNaoModulamPonta}{vetor numerico com codigos dos sitemas que nao modulam na ponta}

\item{sistemasNaoModulamMedia}{vetor numerico com codigos dos sitemas que nao modulam na media}

\item{codTucurui}{codigo atribuido para a usina de Tucurui ###############################}

\item{cotaLimiteTucurui}{valor da cota da usina de Tucurui em metros}

\item{geracaoLimiteTucurui}{valor da geracao limite da usina de Tucurui}
}
\value{
\code{mensagem} vetor de caracteres com a mensagem de sucesso de gravacao na base
}
\description{
Faz a leitura dos arquivos do NEWAVE e carrega todos os dados necessarios na base SQLite para execucao do Balanco de Ponto.
Usa como referencia para a identificacao dos arquivos as posicoes definidas no Manual do Usuario do
Modelo Estrategico de geracao hidrotermica a subsistemas equivalentes do Projeto NEWAVE versao 25.0.1 de setembro/2019 - paginas 12 - 15
}
