% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadosDistDeficit.R
\name{dadosDistDeficit}
\alias{dadosDistDeficit}
\title{Exporta os dados de distribuicao dos deficits}
\usage{
dadosDistDeficit(baseSQLite, tipoCaso, numeroCaso, codModelo)
}
\arguments{
\item{baseSQLite}{nome e localizacao da base SQLite do balanco de potencia}

\item{tipoCaso}{valor inteiro. 1:PDE; 2:PMO e 3;Garantia Fisica}

\item{numeroCaso}{valor inteiro com o numero do caso}

\item{codModelo}{valor inteiro com o codigo do modelo. 1:NEWAVE; 2:SUISHI}
}
\value{
lt.distDef lista com os dados de distribuicao dos deficits, distribuicao de CVaR e comparacao de distribuicao de deficit com CVaR maximo
}
\description{
Exporta os dados de distribuicao dos deficits
}
