% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadosFatorDispHidro.R
\name{dadosFatorDispHidro}
\alias{dadosFatorDispHidro}
\title{Exporta os dados de fator de disponibilidade hidro}
\usage{
dadosFatorDispHidro(
  baseSQLite,
  tipoCaso,
  numeroCaso,
  codModelo,
  inicioHorizonte,
  fimHorizonte,
  pctCenarios = 5
)
}
\arguments{
\item{baseSQLite}{nome e localizacao da base SQLite do balanco de potencia}

\item{tipoCaso}{valor inteiro. 1:PDE; 2:PMO e 3;Garantia Fisica}

\item{numeroCaso}{valor inteiro com o numero do caso}

\item{codModelo}{valor inteiro com o codigo do modelo. 1:NEWAVE; 2:SUISHI}

\item{inicioHorizonte}{valor numerico do ano de inicio do horizonte para o grafico. Formato: AAAA. Ex: 2020}

\item{fimHorizonte}{valor numerico do ano de fim do horizonte para o grafico. Formato: AAAA. Ex:2029}

\item{pctCenarios}{porcentagem de cenarios para serem considerados no calculo da sobra de disponibilidade}
}
\value{
tib.dadosSobraDispHidro tibble com os dados de fator de disponibilidade hidro
}
\description{
Exporta os dados de fator de disponibilidade hidro
}
