% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadosGraficoCVARSubsistema.R
\name{dadosGraficoCVARSubsistema}
\alias{dadosGraficoCVARSubsistema}
\title{Exporta os dados dos graficos de risco de defict}
\usage{
dadosGraficoCVARSubsistema(
  baseSQLite,
  tipoCaso,
  numeroCaso,
  codModelo,
  inicioHorizonteGrafico,
  fimHorizonteGrafico
)
}
\arguments{
\item{baseSQLite}{nome e localizacao da base SQLite do balanco de potencia}

\item{tipoCaso}{valor inteiro. 1:PDE; 2:PMO e 3;Garantia Fisica}

\item{numeroCaso}{valor inteiro com o numero do caso}

\item{codModelo}{valor inteiro com o codigo do modelo. 1:NEWAVE; 2:SUISHI}

\item{inicioHorizonteGrafico}{valor numerico do ano de inicio do horizonte para o grafico. Formato: AAAA. Ex: 2020}

\item{fimHorizonteGrafico}{valor numerico do ano de fim do horizonte para o grafico. Formato: AAAA. Ex:2029}
}
\value{
tib.resultadosCvarMes tibble com os dados do grafico de CVaR por subsistema
}
\description{
Exporta os dados dos graficos de risco de defict de um caso especifico
}
