% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadosGraficoRiscoDeficit.R
\name{dadosGraficoRiscoDeficit}
\alias{dadosGraficoRiscoDeficit}
\title{Exporta os dados dos graficos de risco de defict}
\usage{
dadosGraficoRiscoDeficit(baseSQLite, tipoCaso, numeroCaso, codModelo,
  inicioHorizonteGrafico, fimHorizonteGrafico,
  tituloGrafico = paste0("Risco de Déficit - Caso ", numeroCaso))
}
\arguments{
\item{baseSQLite}{nome e localizacao da base SQLite do balanco de ponta}

\item{tipoCaso}{valor inteiro. 1:PDE; 2:PMO e 3;Garantia Fisica}

\item{numeroCaso}{valor inteiro com o numero do caso}

\item{codModelo}{valor inteiro com o codigo do modelo. 1:NEWAVE; 2:SUISHI}

\item{tituloGrafico}{vetor de caracteres com o titulo do grafico de risco - Nao obrigatorio - valor padrao com numero do caso}

\item{inicioHorizonte}{valor numerico do ano de inicio do horizonte para o grafico. Formato: AAAA. Ex: 2020}

\item{fimHorizonte}{valor numerico do ano de fim do horizonte para o grafico. Formato: AAAA. Ex:2029}
}
\value{
tib.resultadosRisco tibble com os dados do grafico de risco
}
\description{
Exporta os dados dos graficos de risco de defict de um caso especifico
}
