% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadosGraficoRequisitoPot.R
\name{dadosRequisitoPot}
\alias{dadosRequisitoPot}
\title{Exporta os dados do grafico de requisitos de potencia}
\usage{
dadosRequisitoPot(
  baseSQLite,
  tipoCaso,
  numeroCaso,
  codModelo,
  inicioHorizonteGrafico,
  fimHorizonteGrafico
)
}
\arguments{
\item{baseSQLite}{nome e localizacao da base SQLite do balanco de potencia}

\item{tipoCaso}{valor inteiro. 1:PDE; 2:PMO e 3;Garantia Fisica}

\item{numeroCaso}{valor inteiro com o numero do caso}

\item{codModelo}{valor inteiro com o codigo do modelo. 1:NEWAVE; 2:SUISHI}

\item{inicioHorizonte}{valor numerico do ano de inicio do horizonte para o grafico. Formato: AAAA. Ex: 2020}

\item{fimHorizonte}{valor numerico do ano de fim do horizonte para o grafico. Formato: AAAA. Ex:2029}
}
\value{
tib.requisitosPot tibble com os dados do grafico de Requisitos de Potencia
}
\description{
Exporta os dados do grafico de requisitos de potencia
}
