% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficosGF.R
\name{graficosGF}
\alias{graficosGF}
\title{Exibe graficos de CVaR com criterios de GF}
\usage{
graficosGF(
  baseSQLite,
  tipoCaso,
  numeroCaso,
  codModelo,
  tipoGrafico,
  tituloGraficoCVARMes = paste0("Profundidade de Déficit - CVaR Mensal 5\% - Caso ",
    numeroCaso),
  tituloGraficoVARMes = paste0("Profundidade de Déficit - VaR Mensal 5\% - Caso ",
    numeroCaso),
  tituloGraficoVARAno = paste0("Profundidade de Déficit - VaR Anual 5\% - Caso ",
    numeroCaso)
)
}
\arguments{
\item{baseSQLite}{nome e localizacao da base SQLite do balanco de potencia}

\item{tipoCaso}{valor inteiro. 1:PDE; 2:PMO e 3;Garantia Fisica}

\item{numeroCaso}{valor inteiro com o numero do caso}

\item{codModelo}{valor inteiro com o codigo do modelo. 1:NEWAVE; 2:SUISHI}

\item{tipoGrafico}{valor numerico identificando o tipo de grafico. 10: CVAR mensal; 11: VAR mensal; 12: VAR anual}

\item{tituloGraficoCVARMes}{vetor de caracteres com o titulo do grafico de CVAR mensal - Nao obrigatorio - valor padrao com numero do caso}

\item{tituloGraficoVARMes}{vetor de caracteres com o titulo do grafico de VAR mensal - Nao obrigatorio - valor padrao com numero do caso}

\item{tituloGraficoVARAno}{vetor de caracteres com o titulo do grafico de VAR anual - Nao obrigatorio - valor padrao com numero do caso}
}
\value{
objeto do tipo plotly
}
\description{
Monta graficos de Profundidade de Deficit por CVAR de um caso especifico com criterios de GF
}
