% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravaDadosArmazenamentoBDBP.R
\name{gravacaoDadosArmazenamentoBDBP}
\alias{gravacaoDadosArmazenamentoBDBP}
\title{Gravacao dos dados de disponilidade das fontes de tecnologia de armazenamento}
\usage{
gravacaoDadosArmazenamentoBDBP(
  pastaCaso,
  conexao,
  tipoCaso,
  numeroCaso,
  codModelo,
  anoMesInicioMDI,
  anoMesFimMDI
)
}
\arguments{
\item{pastaCaso}{caracter com a localizacao dos arquivos NEWAVE e auxliares do BP.}

\item{conexao}{caracter com a conexao com o Banco de Dados do Balanco de Potencia.}

\item{tipoCaso}{caracter com o tipo de caso simulado. [1]=PDE [2]=PMO [3]=GF.}

\item{numeroCaso}{caracter com o numero do caso, definido pelo usuario.}

\item{codModelo}{caracter com a definicao do modelo utilizado. [1]=Newave [2]=Suishi.}

\item{anoMesInicioMDI}{caracter com o ano e mes do inicio da simulacao do MDI.}

\item{anoMesFimMDI}{caracter com o ano e mes do final da simulacao do MDI.}
}
\value{
\code{mensagem} vetor de caracteres com a mensagem de sucesso de gravacao na tabela BPO_A32_DISPONIBILIDADE_ARMAZENAMENTO
}
\description{
Faz a gravacao dos dados das fontes de tecnologia de armazenamento no banco de dados do Balanco de Potencia (BDBP)
Os dados sao gravados na tabela BPO_A32_DISPONIBILIDADE_ARMAZENAMENTO do BDBP.
}
\examples{
\dontrun{
gravacaoDadosArmazenamentoBDBP("C:/PDE2027_Caso080", conexao, 1, 80, 1, 201901, 203312)}

}
