% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravaDadosTermeletricasBDBP.R
\name{gravacaoDadosTermeletricasBDBP}
\alias{gravacaoDadosTermeletricasBDBP}
\title{Gravacao dos dados das usinas termeletricas ao longo do horizonte de simulacao}
\usage{
gravacaoDadosTermeletricasBDBP(
  pastaCaso,
  pastaSaidas,
  conexao,
  tipoCaso,
  numeroCaso,
  codModelo
)
}
\arguments{
\item{pastaCaso}{caracter com a localizacao dos arquivos NEWAVE e auxliares do BP.}

\item{pastaSaidas}{localizacao dos arquivos de saida do modulo NWLISTOP}

\item{conexao}{caracter com a conexao com o Banco de Dados do Balanco de Potencia.}

\item{tipoCaso}{caracter com o tipo de caso simulado. [1]=PDE [2]=PMO [3]=GF.}

\item{numeroCaso}{caracter com o numero do caso, definido pelo usuario.}

\item{codModelo}{caracter com a definicao do modelo utilizado. [1]=Newave [2]=Suishi.}
}
\value{
\code{mensagem} vetor de caracteres com a mensagem de sucesso de gravacao na tabela BPO_A14_DISPONIBILIDADE_UTE
}
\description{
Faz a gravacao dos dados das usinas termeletricas ao longo do horizonte de simulacao do NEWAVE no banco de dados do Balanco de Potencia (BDBP)
Os dados sao gravados na tabela BPO_A14_DISPONIBILIDADE_UTE do BDBP.
}
\examples{
\dontrun{
gravacaoDadosTermeletricasBDBP("C:/PDE2027_Caso080", "C:/PDE2027_Caso080/nwlistop", 
conexao, 1, 80, 1)
}

}
