% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravaSaidasAnalise.R
\name{gravacaoSaidasAnalises}
\alias{gravacaoSaidasAnalises}
\title{Gravacao dos principais dados de saida}
\usage{
gravacaoSaidasAnalises(
  baseSQLite,
  tipoCaso,
  numeroCaso,
  codModelo,
  df.dadosGerais
)
}
\arguments{
\item{baseSQLite}{caracter com a localizacao da base de dados do BP.}

\item{tipoCaso}{valor inteiro. 1:PDE; 2:PMO e 3;Garantia Fisica}

\item{numeroCaso}{valor inteiro com o numero do caso}

\item{codModelo}{valor inteiro com o codigo do modelo. 1:NEWAVE}

\item{df.dadosGerais}{data frame com os dados gerais do caso utilizado}
}
\value{
\code{mensagem} vetor de caracteres com a mensagem de sucesso de gravacao dos dados de saida de analise
}
\description{
Faz a gravacao dos dados das principais saidas de analise do balanCo de ponta em arquivo excel e no banco de dados do Balanco de Potencia (BDBP).
Os dados sao gravados nas tabelas a partir da BPO_A22 do BDBP
}
\examples{
\dontrun{
gravacaoSaidasAnalises("C:/PDE2027_Caso080.sqlite3", 1, 80, 1)
}

}
